pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safe_code_env"
version = "0.1.0"
description = "Safe Code Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastembed>=0.2.0",
    "numpy>=1.24.0",
    "fastapi>=0.104.0",
    "pytest>=8.0.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "email-validator>=2.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "safe_code_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["safe_code_env", "safe_code_env.server"]
package-dir = { "safe_code_env" = ".", "safe_code_env.server" = "server" }

[tool.setuptools.package-data]
"safe_code_env.server" = ["tasks/**/*", "base_codebase/**/*"]
